Additional fixes for CUDA: move spir-v 1.3 test cases to separate files
